requirements.txt
_libgcc_mutex==0.1
_openmp_mutex==4.5
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.5.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
blas==1.0
brotli-python==1.0.9
bzip2==1.0.8
ca-certificates==2024.9.24
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cudatoolkit==11.3.1
cymem==2.0.6
cython-blis==0.7.9
dataclasses-json==0.6.7
datasets==2.16.1
decorator==5.1.1
dill==0.3.7
distro==1.9.0
evaluate==0.4.3
exceptiongroup==1.2.2
executing==2.1.0
faiss==1.8.0
faiss-cpu==1.8.0
ffmpeg==4.3
filelock==3.16.1
freetype==2.12.1
frozenlist==1.5.0
fsspec==2023.10.0
gmp==6.2.1
gnutls==3.6.15
google==3.0.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.7
iniconfig==2.0.0
intel-openmp==2023.1.0
ipython==8.12.3
jedi==0.19.2
jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
jpeg==9e
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
lame==3.100
langchain==0.1.0
langchain-community==0.0.9
langchain-core==0.1.7
langcodes==3.3.0
langsmith==0.0.92
lcms2==2.12
ld_impl_linux-64==2.40
lerc==3.0
libblas==3.9.0
libdeflate==1.17
libedit==3.1.20230828
libfaiss==1.8.0
libffi==3.4.4
libgcc==14.2.0
libgcc-ng==14.2.0
libgfortran==14.2.0
libgfortran-ng==14.2.0
libgfortran5==14.2.0
libgomp==14.2.0
libiconv==1.16
libidn2==2.3.4
liblapack==3.9.0
libpng==1.6.39
libstdcxx==14.2.0
libstdcxx-ng==14.2.0
libtasn1==4.19.0
libtiff==4.5.1
libunistring==0.9.10
libwebp-base==1.3.2
lz4-c==1.9.4
markdown-it-py==2.2.0
markupsafe==2.1.3
marshmallow==3.22.0
matplotlib-inline==0.1.7
mdurl==0.1.0
mkl==2023.1.0
mkl-service==2.4.0
mkl_fft==1.3.8
mkl_random==1.2.4
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.15
murmurhash==1.0.7
mypy-extensions==1.0.0
ncurses==6.4
nettle==3.7.3
networkx==3.1
nltk==3.9.1
numpy==1.24.3
numpy-base==1.24.3
openai==1.6.1
openh264==2.1.1
openjpeg==2.5.2
openssl==3.4.0
packaging==23.2
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip==24.2
pluggy==1.5.0
preshed==3.0.6
prompt-toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==17.0.0
pyarrow-hotfix==0.6
pydantic==2.10.0
pydantic-core==2.27.0
pygments==2.18.0
pysocks==1.7.1
pytest==8.3.3
python==3.8.20
python-dateutil==2.9.0.post0
python_abi==3.8
pytorch==1.12.1
pytorch-mutex==1.0
pytz==2024.2
pyyaml==6.0.2
readline==8.2
regex==2024.11.6
requests==2.32.3
rich==13.7.1
rouge-score==0.1.2
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==3.2.1
sentencepiece==0.2.0
setuptools==75.1.0
shellingham==1.5.0
six==1.16.0
smart_open==5.2.1
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlalchemy==2.0.36
sqlite==3.45.3
srsly==2.4.8
stack-data==0.6.3
sympy==1.13.3
tbb==2021.8.0
tenacity==8.5.0
thinc==8.2.2
threadpoolctl==3.5.0
tiktoken==0.7.0
tk==8.6.14
tokenizers==0.20.3
tomli==2.1.0
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.67.0
traitlets==5.14.3
transformers==4.46.3
typer==0.9.0
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
urllib3==2.2.3
wasabi==0.9.1
wcwidth==0.2.13
weasel==0.3.4
wheel==0.44.0
xxhash==3.5.0
xz==5.4.6
yarl==1.15.2
zlib==1.2.13
zstd==1.5.6